+++ qpbenchmark/__init__.py
@@ -12,6 +12,7 @@ from .results import Results
 from .run import run
 from .spdlog import logging
 from .test_set import TestSet
+from .tolerance import Tolerance
 from .version import get_version
 
 __version__ = get_version()
@@ -21,6 +22,7 @@ __all__ = [
     "Report",
     "Results",
     "TestSet",
+    "Tolerance",
     "logging",
     "run",
 ]